Core: move Placement out of the raster backend and drop its glyph-only twin
